+++ ats-pkg/src/Language/ATS/Package.hs
@@ -6,6 +6,8 @@ module Language.ATS.Package ( buildAll
                             -- * Ecosystem functionality
                             , displayList
                             , atspkgVersion
+                            -- * Functions involving the compiler
+                            , packageCompiler
                             -- * Types
                             , Version (..)
                             , Pkg (..)