+++ ats-pkg/src/Language/ATS/Package/Build.hs
@@ -271,7 +271,9 @@ pkgHome cc' = liftIO $ getAppUserDataDirectory ("atspkg" </> ccToDir cc')
 
 -- | The directory that will be @PATSHOME@.
 patsHomeAtsPkg :: MonadIO m => Version -> m String
-patsHomeAtsPkg v = fmap (</> show v) (pkgHome (GCC Nothing Nothing))
+patsHomeAtsPkg v = fmap (</> vs) (pkgHome (GCC Nothing Nothing))
+    where vs = show v
+          -- gmp = if v >= Version [0,3,13] then "gmp-" else ""
 
 home' :: MonadIO m => Version -- ^ Compiler version
                    -> Version -- ^ Library version