+++ shake-ats/CHANGELOG.md
@@ -1,5 +1,9 @@
 # shake-ats
 
+## 1.10.2.1
+
+  * Use oracle to track `HsCompiler` in rules
+
 ## 1.10.2.0
 
   * `genATS` takes `cpphs` parameter again